+++ services/user-service/user-service-core/src/main/java/com/kabunx/erp/mapper/AdminMapper.java
@@ -10,6 +10,6 @@ import java.util.List;
 
 @Mapper
 public interface AdminMapper extends BaseMapper<AdminDO> {
-    AdminVO findByUserId(Integer userId);
+    AdminVO findByUserId(Long userId);
     List<AdminVO> selectWithUser(IPage<AdminVO> page);
 }
